+++ crates/vizia_core/src/views/mod.rs
@@ -19,7 +19,7 @@ pub mod normalized_map;
 mod notification;
 mod picklist;
 mod popup;
-mod radio_buttons;
+mod radio;
 mod rating;
 mod scrollbar;
 mod scrollview;
@@ -54,7 +54,7 @@ pub use menu::*;
 pub use notification::Notification;
 pub use picklist::PickList;
 pub use popup::{Popup, PopupData, PopupEvent};
-pub use radio_buttons::RadioButton;
+pub use radio::RadioButton;
 pub use rating::Rating;
 pub use scrollbar::Scrollbar;
 pub use scrollview::{ScrollData, ScrollEvent, ScrollView};
